Bug 1970922: Increase size of ObjectFlags r=jandem
This does not increase memory usage. The layout of Shape is:
```
uint32_t immutableFlags;
ObjectFlags objectFlags_;
ShapeCachePtr cache_;
```

Prior to this patch we had a 2-byte hole following objectFlags_ (on both 32-bit and 64-bit). This patch allows us to use those bytes.

Differential Revision: https://phabricator.services.mozilla.com/D252931

+++ js/src/vm/ObjectFlags.h
@@ -19,7 +19,7 @@ namespace js {
 //
 // If you add a new flag here, please add appropriate code to JSObject::dump to
 // dump it as part of the object representation.
-enum class ObjectFlag : uint16_t {
+enum class ObjectFlag : uint32_t {
   IsUsedAsPrototype = 1 << 0,
   NotExtensible = 1 << 1,
   Indexed = 1 << 2,
